Add loading property to auth state, move accounts api file
