--- client/src/views/Auth/components/Form/api.js
@@ -1,10 +1,0 @@
-import _requests from '@/services/agent';
-
-const prefix = process.env.REACT_APP_ACCOUNTS_ROOT;
-const requests = _requests({ prefix });
-
-const validationApi = {
-  validateFields: values => requests.get('/validate', values)
-};
-
-export default validationApi;
